Reorganizes directory structure and command hierarchy
