--- cmd/deletedeviceservice.go
@@ -1,80 +1,0 @@
-// Copyright © 2019 NAME HERE <EMAIL ADDRESS>
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cmd
-
-import (
-	"fmt"
-	"io/ioutil"
-	"net/http"
-
-	"github.com/spf13/cobra"
-)
-
-// deletedeviceserviceCmd represents the deletedeviceservice command
-var deletedeviceserviceCmd = &cobra.Command{
-	Use:   "deletedeviceservice",
-	Short: "Delete device service by name",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println(args[0])
-
-		client := &http.Client{}
-
-		// Create request
-		req, err := http.NewRequest("DELETE", "http://localhost:48081/api/v1/deviceservice/name/"+args[0], nil)
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		// Fetch Request
-		resp, err := client.Do(req)
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-		defer resp.Body.Close()
-
-		respBody, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		// Display Results
-		fmt.Println("response Status : ", resp.Status)
-		fmt.Println("response Headers : ", resp.Header)
-		fmt.Println("response Body : ", string(respBody))
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(deletedeviceserviceCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// deletedeviceserviceCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// deletedeviceserviceCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
