+++ cmd/profile/profile.go
@@ -1,0 +1,67 @@
+// Copyright © 2019 NAME HERE <EMAIL ADDRESS>
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package profile
+
+import (
+	listprofile "github.com/edgexfoundry/edgex-cli/cmd/profile/list"
+	rmprofile "github.com/edgexfoundry/edgex-cli/cmd/profile/rm"
+	"github.com/spf13/cobra"
+)
+
+// NewCommand returns the profile command
+func NewCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "profile [option]",
+		Short: "A brief description of your command",
+		Long: `A longer description that spans multiple lines and likely contains examples
+	and usage of using your command. For example:
+
+	Cobra is a CLI library for Go that empowers applications.
+	This application is a tool to generate the needed files
+	to quickly create a Cobra application.`,
+	}
+
+	cmd.AddCommand(rmprofile.NewCommand())
+	cmd.AddCommand(listprofile.NewCommand())
+	return cmd
+}
+
+// var listCmd = &cobra.Command{
+// 	Use:   "list",
+// 	Short: "A brief description of your command",
+// 	Long: `A longer description that spans multiple lines and likely contains examples
+// and usage of using your command. For example:
+
+// Cobra is a CLI library for Go that empowers applications.
+// This application is a tool to generate the needed files
+// to quickly create a Cobra application.`,
+// 	Run: func(cmd *cobra.Command, args []string) {
+// 		fmt.Println("list called")
+// 	},
+// }
+
+// var rmCmd = &cobra.Command{
+// 	Use:   "rm",
+// 	Short: "A brief description of your command",
+// 	Long: `A longer description that spans multiple lines and likely contains examples
+// and usage of using your command. For example:
+
+// Cobra is a CLI library for Go that empowers applications.
+// This application is a tool to generate the needed files
+// to quickly create a Cobra application.`,
+// 	Run: func(cmd *cobra.Command, args []string) {
+// 		fmt.Println("rm called")
+// 	},
+// }
