+++ cmd/root.go
@@ -1,4 +1,4 @@
-// Copyright © 2019 NAME HERE <EMAIL ADDRESS>
+// Copyright © 2019 VMware, INC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,72 +18,87 @@ import (
 	"fmt"
 	"os"
 
-	homedir "github.com/mitchellh/go-homedir"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
+
+	"github.com/edgexfoundry/edgex-cli/cmd/device"
+	"github.com/edgexfoundry/edgex-cli/cmd/event"
+	"github.com/edgexfoundry/edgex-cli/cmd/profile"
 )
 
 var cfgFile string
 
 // rootCmd represents the base command when called without any subcommands
-var rootCmd = &cobra.Command{
-	Use:   "edgex-cli",
-	Short: "EdgeX command line interface",
-	Long: `
+func NewCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "edgex-cli",
+		Short: "EdgeX command line interface",
+		Long: `
 EdgeX Command Line Interface
 
 https://www.edgexfoundry.org/
 	`,
 
-	// Uncomment the following line if your bare application
-	// has an action associated with it:
-	//	Run: func(cmd *cobra.Command, args []string) { },
+		// Uncomment the following line if your bare application
+		// has an action associated with it:
+		//	Run: func(cmd *cobra.Command, args []string) { },
+	}
+
+	// Add all subcommands below:
+	cmd.AddCommand(device.NewCommand())
+	cmd.AddCommand(profile.NewCommand())
+	cmd.AddCommand(event.NewCommand())
+
+	return cmd
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
-	if err := rootCmd.Execute(); err != nil {
+	if err := NewCommand().Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
 }
 
-func init() {
-	cobra.OnInitialize(initConfig)
-
-	// Here you will define your flags and configuration settings.
-	// Cobra supports persistent flags, which, if defined here,
-	// will be global for your application.
-	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.edgex-cli.yaml)")
-
-	// Cobra also supports local flags, which will only run
-	// when this action is called directly.
-	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+func Main() {
+	Execute()
 }
 
-// initConfig reads in config file and ENV variables if set.
-func initConfig() {
-	if cfgFile != "" {
-		// Use config file from the flag.
-		viper.SetConfigFile(cfgFile)
-	} else {
-		// Find home directory.
-		home, err := homedir.Dir()
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
-		}
+// func init() {
+// 	cobra.OnInitialize(initConfig)
 
-		// Search config in home directory with name ".edgex-cli" (without extension).
-		viper.AddConfigPath(home)
-		viper.SetConfigName(".edgex-cli")
-	}
+// 	// Here you will define your flags and configuration settings.
+// 	// Cobra supports persistent flags, which, if defined here,
+// 	// will be global for your application.
+// 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.edgex-cli.yaml)")
 
-	viper.AutomaticEnv() // read in environment variables that match
+// 	// Cobra also supports local flags, which will only run
+// 	// when this action is called directly.
+// 	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+// }
 
-	// If a config file is found, read it in.
-	if err := viper.ReadInConfig(); err == nil {
-		fmt.Println("Using config file:", viper.ConfigFileUsed())
-	}
-}
+// // initConfig reads in config file and ENV variables if set.
+// func initConfig() {
+// 	if cfgFile != "" {
+// 		// Use config file from the flag.
+// 		viper.SetConfigFile(cfgFile)
+// 	} else {
+// 		// Find home directory.
+// 		home, err := homedir.Dir()
+// 		if err != nil {
+// 			fmt.Println(err)
+// 			os.Exit(1)
+// 		}
+
+// 		// Search config in home directory with name ".edgex-cli" (without extension).
+// 		viper.AddConfigPath(home)
+// 		viper.SetConfigName(".edgex-cli")
+// 	}
+
+// 	viper.AutomaticEnv() // read in environment variables that match
+
+// 	// If a config file is found, read it in.
+// 	if err := viper.ReadInConfig(); err == nil {
+// 		fmt.Println("Using config file:", viper.ConfigFileUsed())
+// 	}
+// }
